Added options to map object. Use instead of javascript interop parameters. In my application, I like to set the max bounds so that the user cannot zoom out to see the world repeated because markers and other layers only show in one spot and are not repeated.

+++ BlazorLeaflet/BlazorLeaflet/Models/GridLayer.cs
@@ -28,7 +28,7 @@ namespace BlazorLeaflet.Models
         /// <summary>
         /// If set, tiles will only be loaded inside the set.
         /// </summary>
-        public Tuple<double, double> Bounds { get; set; }
+        public Tuple<float, float> Bounds { get; set; }
 
     }
 }
